funding detail, list
