+++ idolsns/src/main/java/com/kh/idolsns/dto/FundPostDto.java
@@ -10,6 +10,7 @@ public class FundPostDto {
 	private Long postNo;
 	private String memberId;
 	private String fundTitle;
+	private String fundShortTitle;
 	private Date postStart;
 	private Date postEnd;
 	private int fundGoal;
