+++ idolsns/src/main/java/com/kh/idolsns/dto/FundPostImageDto.java
@@ -10,6 +10,7 @@ public class FundPostImageDto {
 	private Long postNo;
 	private String memberId;
 	private String fundTitle;
+	private String fundShortTitle;
 	private Date postStart;
 	private Date postEnd;
 	private Date postTime;
@@ -17,7 +18,7 @@ public class FundPostImageDto {
 	private int fundSponsorCount;
 	private String fundState;
 	private String postType;
-	private String postContent; 
+	private String postContent;
 	private Integer attachmentNo; // 없는 경우 null이 나오도록 처리
