+++ idolsns/src/main/java/com/kh/idolsns/restcontroller/FundRestController.java
@@ -1,5 +1,8 @@
 package com.kh.idolsns.restcontroller;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,7 +11,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.kh.idolsns.dto.FundDto;
+import com.kh.idolsns.dto.FundPostImageDto;
+import com.kh.idolsns.dto.PostImageDto;
+import com.kh.idolsns.repo.FundPostImageRepo;
 import com.kh.idolsns.repo.FundRepo;
+import com.kh.idolsns.vo.FundDetailVO;
 
 @CrossOrigin
 @RestController
@@ -18,21 +25,53 @@ public class FundRestController {
 	@Autowired
 	private FundRepo fundRepo;
 	
-	// 후원하기 
-//	@PostMapping("/")
-//	public void add(
-//			@ParameterObject
-//			@RequestBody FundDto fundDto
-//			) {
-//		Long fundNo = fundRepo.sequence();
-//		fundDto.setFundNo(fundNo);
-//		// memberId는 header.jsp에서 가져와서 사용
-//		fundRepo.insert(fundDto);
-//		
-//	}
+	@Autowired
+	private FundPostImageRepo fundPostImageRepo;
+	
+	// 펀딩게시물 목록 조회
+	@GetMapping("/")
+	public List<FundPostImageDto> fundPostList(){
+		return fundPostImageRepo.selectList();
+	}
+	
+	// 펀딩상세 
+	@GetMapping("/{postNo}")
+	public FundPostImageDto detail(@PathVariable Long postNo) {
+		FundDetailVO vo = fundPostImageRepo.selectOne(postNo);
+		FundPostImageDto fundPostImageDto = vo.getFundPostImageDto();
+//		System.out.println(fundPostImageDto);
+		return fundPostImageDto;
+	}
+	
+	// 상세이미지 attachmentNos 
+	@GetMapping("/attaches/{postNo}")
+	public List<Integer> list(@PathVariable Long postNo) {
+		List<PostImageDto> list = fundPostImageRepo.selectAttachList(postNo);
+		List<Integer> attachList = new ArrayList<>();
+		
+		for(PostImageDto dto : list) {
+			attachList.add(dto.getAttachmentNo());
+		}
+		
+		return attachList;
+	}
+	
+	
+	// 후원한 total금액 
+	@GetMapping("/fundtotal/{postNo}")
+	public int fundTotal(@PathVariable Long postNo){
+		List<FundDto> list = fundPostImageRepo.selectFundList(postNo);
+		
+		int fundTotal = 0;
+		for(FundDto dto : list) {
+			fundTotal += dto.getFundPrice();
+		}
+		
+		return fundTotal;
+	}
 	
 
-	  @GetMapping("/{fundNo}")
+	  @GetMapping("/order/{fundNo}")
 	  public FundDto getFund(@PathVariable long fundNo) {
 	    FundDto fundDto = fundRepo.find(fundNo);
 	    return fundDto;
