+++ idolsns/src/main/resources/mybatis/mapper/fundpost-mapper.xml
@@ -11,11 +11,11 @@
 	<!-- C 펀딩 게시물 등록 -->
 	<insert id="add" parameterType="FundPostDto">
 		insert into fund_post(
-					post_no, member_id, fund_title, post_start, post_end,
+					post_no, member_id, fund_title, fund_short_title, post_start, post_end,
 					fund_goal, fund_sponsor_count, fund_state
 					)
 		values(
-		#{postNo}, #{memberId}, #{fundTitle}, #{postStart}, 
+		#{postNo}, #{memberId}, #{fundTitle}, #{fundShortTitle}, #{postStart}, 
 		#{postEnd}, #{fundGoal}, #{fundSponsorCount}, '진행중'
 		)
 	</insert>
