+++ idolsns/src/main/resources/mybatis/mapper/fundpostinte-mapper.xml
@@ -5,7 +5,7 @@
     PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
     "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
     
-<mapper namespace="fundpostlist">
+<mapper namespace="fundpostinte">
 
 	<resultMap type="FundDetailVO" id="fdvo">
 		<!-- FundPostImageDto (fund_post + post + fund_main_image) -->
@@ -35,7 +35,7 @@
 	</resultMap>
 	
 	<select id="fundPostByPostNo" resultMap="fdvo" parameterType="Long">
-		select * from fundpost_list 
+		select * from fundpostimage_view
 		where post_type='펀딩' and post_no=#{postNo} 
 	</select>
 	
@@ -49,7 +49,7 @@
 	</select>
 	
 	<select id="list" resultType="FundPostImageDto">
-		select * from fundpost_list where post_type='펀딩'
+		select * from fundpostimage_view where post_type='펀딩'
 		order by post_time desc
 	</select>
